# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/employees": {
      "get": {
        "operationId": "EmployeesController_findAll",
        "parameters": [
          {
            "name": "view",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sort",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "departmentId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "employees"
        ]
      },
      "post": {
        "operationId": "EmployeesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEmployeeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "employees"
        ]
      }
    },
    "/api/v1/employees/{id}": {
      "get": {
        "operationId": "EmployeesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "employees"
        ]
      },
      "patch": {
        "operationId": "EmployeesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEmployeeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "employees"
        ]
      },
      "delete": {
        "operationId": "EmployeesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "employees"
        ]
      }
    },
    "/api/v1/employees/{id}/contracts": {
      "get": {
        "operationId": "EmployeesController_getContracts",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "employees"
        ]
      }
    },
    "/api/v1/employees/{id}/attendance": {
      "get": {
        "operationId": "EmployeesController_getAttendance",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "employees"
        ]
      }
    },
    "/api/v1/employees/{id}/time-off": {
      "get": {
        "operationId": "EmployeesController_getTimeOff",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "employees"
        ]
      }
    },
    "/api/v1/employees/{id}/timeline": {
      "get": {
        "operationId": "EmployeesController_getTimeline",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "employees"
        ]
      }
    },
...
    "/api/v1/attendance/widget/today": {
      "get": {
        "operationId": "AttendanceController_getTodayWidget",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "attendance"
        ]
      }
    },
    "/api/v1/attendance": {
      "get": {
        "operationId": "AttendanceController_findAll",
        "parameters": [
          {
            "name": "employeeId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "dateFrom",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "dateTo",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "attendance"
        ]
      }
    },
    "/api/v1/attendance/{id}": {
      "get": {
        "operationId": "AttendanceController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "attendance"
        ]
      },
      "patch": {
        "operationId": "AttendanceController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAttendanceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "attendance"
        ]
      }
    },
    "/api/v1/attendance/check-in": {
      "post": {
        "operationId": "AttendanceController_checkIn",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckInDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "attendance"
        ]
      }
    },
    "/api/v1/attendance/{id}/check-out": {
      "post": {
        "operationId": "AttendanceController_checkOut",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "attendance"
        ]
      }
    },
...
  },
  "info": {
    "title": "PeoplePay360 API",
    "description": "HR & Payroll Management System API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateEmployeeDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "John Doe"
          },
          "workEmail": {
            "type": "string",
            "example": "john.doe@example.com"
          },
          "jobPosition": {
            "type": "string",
            "example": "Software Engineer"
          },
          "departmentId": {
            "type": "string",
            "example": "uuid-of-department"
          },
          "managerId": {
            "type": "string",
            "example": "uuid-of-manager"
          },
          "workingScheduleId": {
            "type": "string",
            "example": "uuid-of-working-schedule"
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "INACTIVE",
              "ON_LEAVE",
              "TERMINATED"
            ]
          },
          "avatarUrl": {
            "type": "string",
            "example": "https://example.com/avatar.jpg"
          },
          "phone": {
            "type": "string",
            "example": "+1234567890"
          },
          "employeeType": {
            "type": "string",
            "example": "Full-time"
          },
          "bankAccount": {
            "type": "string",
            "example": "1234567890"
          },
          "bankIfsc": {
            "type": "string",
            "example": "SBIN0001234"
          }
        },
        "required": [
          "name",
          "workEmail"
        ]
      },
      "UpdateEmployeeDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "John Doe"
          },
          "workEmail": {
            "type": "string",
            "example": "john.doe@example.com"
          },
          "jobPosition": {
            "type": "string",
            "example": "Software Engineer"
          },
          "departmentId": {
            "type": "string",
            "example": "uuid-of-department"
          },
          "managerId": {
            "type": "string",
            "example": "uuid-of-manager"
          },
          "workingScheduleId": {
            "type": "string",
            "example": "uuid-of-working-schedule"
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "INACTIVE",
              "ON_LEAVE",
              "TERMINATED"
            ]
          },
          "avatarUrl": {
            "type": "string",
            "example": "https://example.com/avatar.jpg"
          },
          "phone": {
            "type": "string",
            "example": "+1234567890"
          },
          "employeeType": {
            "type": "string",
            "example": "Full-time"
          },
          "bankAccount": {
            "type": "string",
            "example": "1234567890"
          },
          "bankIfsc": {
            "type": "string",
            "example": "SBIN0001234"
          }
        }
      },
...
      "CheckInDto": {
        "type": "object",
        "properties": {
          "employeeId": {
            "type": "string",
            "example": "uuid-of-employee"
          }
        }
      },
      "UpdateAttendanceDto": {
        "type": "object",
        "properties": {
          "checkOut": {
            "type": "string",
            "example": "2024-01-01T18:00:00Z"
          },
          "status": {
            "type": "string",
            "enum": [
              "PRESENT",
              "LATE",
              "ABSENT",
              "OVERTIME",
              "MISSING_CHECKOUT",
              "MANUALLY_EDITED"
            ]
          },
          "notes": {
            "type": "string",
            "example": "Manual correction - system error"
          }
        }
      },
...
    }
  }
}